Add files via upload
CSS   stands   for   Cascading   Style Sheets. 
CSS   describes   how   HTML elements are to be displayed on screen, paper,  or  in  other  media. 
It  can  control  the  layout of multiple web pages all at once.
CSS allows you to control the look and feel of  several  pages  by  changing  a  single source.

A CSS rule set contains
a selector and
a declaration block.
 Types of CSS 

There     are     three     methods     of implementing styling

 Inline CSS :- It  uses  the  style attribute   in   the   HTML   start   tag. Inline CSS is used to apply CSS on a single line or element.
For example : <p style="color:blue">Hello CSS</p>

2.Embedded  stylesheet  or  internal CSS : This is used to apply CSS on a  single  document  or  page.  It  can affect all the elements of the page. It is written inside the style tag within head section of html.For example :   <style> h1{color: Red;}</style>

3. External CSS:   The  external style sheet is generally used when you want  to  make  changes  on  multiple pages. It uses the <link> tag on every page and the <link> tag should be put inside the head section. The  external  css file should not contain any HTML tags ,open with any text editor  it saved with .css extension.
For example :   Style.css
h1{color:navy;margin-left:20px}
 <link  rel="stylesheet" href="style.css"> 

1)CSS Id Selector

The   Id   selector   selects   the   id attribute   of   an   HTML   element   to select   a   specific   element.  
 An   id   is  always  unique  within  the  page  so  it  is unique  element.  It  is  written  with  the hash  character(#),  followed  by  the  id name.
2) CSS Class Selector
The   class   selector   selects   HTML elements  with  a  specific  class  attribute.
 It  is  used  with  a  period  character  '.' (full stop symbol) followed by the class name. 
The Class selector is used when you want to change a group of elements within your HTML page. The  class  name  should  not  start with number.
3) Class Selector for specific element
To  specify  only  one  specific  HTML element   should   be   affected   then   you should use the element name with class selector.
4)Universal Selector
The  universal  selector  is  used  as a  wildcard  character.  It  selects  all  the elements on the Webpages.

5)Group Selector
The grouping selector is used to select all   the   elements   with   the   same   style definitions.  It  is  used  to  minimize  the code. 
Commas are used to separate each selector in grouping.
H1,h2,p{ text-align:center;color:blue}
6) Positioning in CSS

CSS  helps  to  position  the    HTML elements. 
The position property is used to   set   position   for   an   element.  
 The element can be positioned using the top, bottom, left and right properties.
7) Display property
The Display property in CSS defines how   the   components   (div,   hyperlink, heading,  etc)  are  going  to  be  placed  on the web page.
 It specifies how the element is displayed. this property is used to define the display of different parts of a web page.
                         
                                 Syntax :       Display : value;

  Where values are :
Inline : It is used to display an element as an inline element.
Block : It is used to display an element as  an  block  element.  It  starts  on  a  new line, and takes up the whole width of the browser window.
Block-inline : This value is very similar to inline element but the difference is that you are able to set the width and height.
None   :   The   element   is   completely removed.

8)Float Property in CSS
Float is a CSS property written in CSS file or directly in the style of an element. The  float  property  defines  the  flow  of content.
Following  are  the  types  of  floating properties :
float  :  left  :  This  keeps  the  element float on left side of the container
float : right : This  keeps the element float on right side of container
float : none :  This is default property    i.e. this shows the element as it is.